requirements.txt
Flask==2.3.3
sentence-transformers==2.2.2
transformers==4.39.3
accelerate==0.18.0
torch==2.1.2
pdfplumber==0.10.3
camelot-py[cv]
azure-storage-file-datalake==12.12.0
gunicorn==20.1.0
python-dotenv==1.0.0
numpy==1.26.4
scikit-learn==1.3.0
